Remove metal driver and move logic back to view
The "driver context" is going to become the "driver".
This is something I realized after implementing cardboard,
which doesn't need a driver like we had before.

Basically for VR Distortion rendering you have a View. Either
VROViewCardboard or VROViewMetal. For scene rendering,
on the other hand, you have a driver you have a driver and its
corresponding substrates.