Fixed Viro Scene Memory Leak Crash
VROController was leaking VROScenes even after we have popped those scenes
and the VRTScene views were destroyed. This is especially bad if we have many
scenes with heavy 360 views that would eat up memory until the app crashes.
Root caused the leak to VROSceneController that held a strong reference
to the hover delegate, that inturn held a strong reference to the controller,
introducing an cyclic reference - leak. Solution is to pass in a weak reference
to the delegate.