VIRO-504 [perf]: switch lights via binding points
Instead of keeping a single UBO for all sets of
lights, create one UBO per set of lights. This way,
switching between sets of lights is as easy as
changing the binding point that the shader's uniform
block is pointing to. Previously, we would rewrite
our single UBO with each set of new lights.