VIRO-508: shader code injection framework
Framework to modify the base shaders used
by the renderer. An entry point and section can
be specified which indicates where to inject the
code. The injected code is placed in a
VROShaderModifier. These modifiers can be
attached to materials, and the injected code will
be compiled at shader construction time.

Still to do: ability to inject values into custom
uniforms, and support for more entry points.

Note the polyline story requires this because we
have to modify the vertex shader to draw polylines.