VIRO-493: Fix onGaze Triggers
Fixed bug where the onGaze callback is triggered when we are gazing near
but not on an object. This results in a seemingly inaccurate larger "hitbox"
due to a bug within our hitTest code, where we are checking with a large margin
of error that favors a "hit situation". As such, reducing the margin of error
to 0.01 resolves the issue - verified with Raj as well.