VIRO-520: fix lighting crash
Crash was caused because the material substrate
is not always created at the time the sort key is
created. In such cases, we were assigning shader
ID 0 for the material. If we then went on the render
that material, we could end up with a random shader
being bound on render time instead of the shader that
should be used. This could cause a crash if the
random floating shader required lights, and our material
didn't bind any lights.

Fixed by making it such that the material substrate
is created on demand when the sort key is updated.

Also removed some uniforms for lights that should have
been removed a long time ago, and removed a the
generic lighting functions from the constant light shaders.