[android]: port renderer, OpenGL driver, audio, animation
Large commit because could not compile any of it until
all of it was ported.

+ ported Renderer package to cross-platform
+ ported Animation package to cross-platform
+ ported Audio package to cross-platform
+ ported OpenGL driver to cross-platform

The key changes were in the way the render delegate
and scene controller were refactored. Now, the renderer
itself uses VRORenderDelegateInternal and
VROSceneControllerInternal. These are C++ classes.
The iOS applications *using* the renderer use
VRORenderDelegate and VROSceneController, which
are Obj-C classes. ViroKit, meanwhile, implements
VRORenderDelegateiOS and VROSceneControlleriOS,
which funnel commands between the Obj-C and C++
classes. This leaves room for Android implementations
as well.

The other significant changes were the introduction of
more platform-specific renderer and utility classes.