[android]: clean up samples and scene controller
Change so that scene controllers no longer need
access to internal variables like the reticle and
frame synchronizer; instead those are acquired
when the controller is attached.

With this change, was able to further decouple
the sample from the scene controller. Now, the
sample serves as a render delegate, and it
creates a scene controller. That scene controller
is then fed into the scene renderer.

Also fixed a JNI error, making it thread-safe.