VIRO-595: animation refactor
Refactor to create execution framework for
animations within the renderer, so that the
bridge only has to handle parsing of animations
from React code, and glueing the animations
to the node they are acting on via the animated
component.

This change introduces the VROExecutableAnimation
interface. It's a composite pattern where
VROAnimationGroup represents a set of property
animations that will run in the same transaction, and
VROAnimationChain represents a list of groups that will
be executed either serially or in parallel.

Each individual property animation is represented
by a VROPropertyAnimation.