VIRO-168 [ovr]: introduce OVR renderer
Introduce the ViroOVRView and the
VROSceneRendererOVR, which implement the
VrView and VROSceneRenderer, respectively.

This is the initial plugin for the OVR system. The
OVR system requires more lifecycle callbacks
than GVR (including all surface holder callbacks),
so these have been plugged through as well. Note
that many of these lifecycle events are no-ops for
GVR.

Currently this gets us to the "Insert your phone into
a Gear VR" screen. The actual scene rendering
still has to be hooked up to the OVR scene renderer,
and the default OVR rendering needs to be removed.