VIRO-168 [ovr]: integrate OVR renderer with VRO
Delete the unused parts of the OVR renderer,
and replace them with hooks into the VRO scene
renderer. With this commit 3D rendering and
backgrounds are working in VR.

This commit also *disables* simultaneous multiview
rendering, as the VRO rendering does not yet
support that extension.

The following remains incomplete:
+ non-VR mode
+ audio
+ queuing to renderer thread
+ frame listening
+ video