VIRO-168 [ovr]: delete multi-thread code
The OVR sample we're basing this off of
included code to integrate with a renderer
where the eyes are rendered on a separate
thread from the distortion renderer. We're
not going to use this, so delete to reduce
complexity.

(Note we still have a rendering thread, which
is distinct from the main/UI thread)