VIRO-168 [ovr]: text rendering
Text rendering was broken because we were
calling the initializeGL() method from the main
thread instead of the rendering thread. This took
some time to figure out (since we weren't on the
rendering thread, the font textures were being
initialized to black). We never ran into this problem
with normal objects because those objects'
textures are not loaded until later, not on initialization.

In the course of figuring this out, added a few
OpenGL checking utilities and macros in VROOpenGL
that can come in handy.

With this commit OVR is feature-complete except
for input handling and pause/resume crash bugs.