VIRO-696: DayDream Controller Integration Part 1
- Renamed VROEventManager to VROControllerInput containing the implementation of
Viro-specific events to be mapped to. This will be then subclassed by platform
specific controllers - VROControllerInputCardboard, VROControllerInputDaydream -
that will implement platform-specific Event handling and map them to the
ViroSpecific ones.
- These platform-specific controllers will be constructed within their
platform specific renderers, but contained as a base class within VRORenderer.
- Expanded VROEventDelegate to include a generic set of controller events
encompassing buttons, touchpad and orientation. This are grouped into
EventSource, descrbing the source of the triggered event, and EventAction,
which decribes the action that the source has taken.
- Added VROControllerPresenter that represents the UI for our input controllers.
Every platform VROControllerInput will create it's corresponding VROControllerPresenter.
- Added Controller_JNI through which developers can access these ControllerPresenters,
say for registering delegates directly, or for custom modification of controller UI.
- Added VROControllerInputs and presenter counterparts for iOS.
- Fixed iOS VROSample to work with our new SceneControllers / Event System.