VIRO-168 [ovr]: detached thread Java invocations
OVR uses a natively-spawned pthread for its
renderer. This thread didn't work with VROPlatformUtil's
VROPlatformCallJavaFunction method because
the pthread, even when attached to the JVM, did
not have the Android class loader attached. In order
to get this to work, added a new findClass method,
VROPlatformFindClass, that takes a java object
argument that we can use to retrieve the correct
class loader. This makes it work.

After implementing this, I realized we didn't need this
at all because we can just use env->GetObjectClass
in VROPlatformCallJavaFunction, so we do that instead.
However, I've left the new FindClass method there
because it's a useful utility function for JNI which we
may need if we ever want to invoke a Java method from
the pthread *without* already having the target object.