VIRO-715: multi-axis animations
We now support parallel animations that
modify multiple axes of a single property.
Previously, if two animations modified different
axes (i.e. rotateX and rotateY), only the last
animation would be processed, and the
previous ones would be clobbered. This was
because we only allowed animating an entire
vector at a time, as opposed to a single axis
of the vector.