[animation]: make material cross-fades better
Cross-fades between opaque materials would
expose the background during the transition
(because we were rendering alpha < 1 on top of
alpha < 1). So instead we now only fade in the
incoming material when transitioning materials.

For transparent materials, we maintain the usual
cross-fade; it looks better, and minimizes 'popping'
at the end when the outgoing material disappears.