Default materials to write/read from depth buffer
Previously, materials defaulted to NOT write/reading
from the depth buffer. This created complications
because Viro components generally will use the depth
buffer, so the bridge kept having to set these to true.

Instead, we should make *not* using the depth buffer
the exception, requiring the developer to use a custom
Material if he/she wants to disable the depth buffer.

The only exception here is the skybox and video spheres;
those we will default to not reading/writing from the depth
buffer, because they are backgrounds.