VIRO-592: hierarchical rendering for flexview
Distance from camera (back to front) rendering
was causing issues inside flex views because
sometimes the background views would be closer
to the camera than their children (because we test
based on the center of each view). To get around
this, introduced a new flag called hierarchicalRendering
to VRONode. When this flag is true, all children
of the node will be rendered hierarchically, based on
their scene graph depth. This system is better suited
to rendering layers of 2D views. 3D objects still use
the existing render order system.