VIRO-433 [renderer]: node-based cameras
Integrates cameras with nodes. Cameras can be
added to nodes, and then a specific node can be
chosen to be the "point of view" for the scene.
This also enables cameras to be animated.

This change includes all the rendering changes, along
with the JNI additions required for the bridge.