VIRO-765: Resolve Android Renderer Threading Issues
- Dispatched GL specific Viro components to be modified on the Renderer Thread.
These are mainly node specific components that are examined / modified in the
render pass.
- Dispatching calls from React-to-Renderer is done at the JNI C++ layer. Because
all of these calls are mainly setters, they can ben done asynchronously.
- Added ability to post onto Android's UI Applicaiton thread via VROPlatformUtil,
where delegate events to be sent back to JS are then posted on the UI thread.
- Modified VideoTextures to be created asynchronously to account for gl calls.
- Moved the direct calling into Native for Camera position to be async instead.
- Moved the check of containsChild from NodeJNI into the Node view layer so
as to remove the requirement of calling into Native; we use the view hiearchy
in android to perform the same check instead.