VIRO-128: refactor iOS material animations
Refactor the iOS material animations to use
a more tightly encapsulated lazy loading
pattern instead of a delegate pattern. This
moves the actual animation logic into the renderer,
whereas previously it was in the React layer
via the delegate.

Also introduces a new VROMaterialAnimation,
because VROPropertyAnimation was not a good
fit for material animations, and enables animating
multiple material indices at once.