VIRO-889,VIRO-890: LifeCycle Crashes for GVR and OVR
- Both gvr and ovr layouts were listening for lifecycle callbacks that
may not have originated from the React activity, causing onPause to
be called on activities that have already paused. Thus fix is to
save a weak reference of the react activity to compare against.

- Fixed Memory-Leak-OVR-LifeCycle acyclic dependency issue where Viro Controls
that had references to the renderer were not being deallocated properly. This
is because even though they were queued to be deallocted on the RenderThread,
the onDestroy Msg required by OVRRenderer to process that queue is only processed
on the destruction of OVRRenderer, of which it's ref is held by the very Viro
Controls that cannot be destroyed. Thus, fix is to trigger an onDestory Msg apart
from the destructor to break that dependency, allowing for proper deallocations
to occur.