VIRO-876: scene navigation opacities stay black
Incoming transitions were reading the opacity to
transition toward from the node itself. The problem
was if those nodes were previously outgoing, that
opacity would be zero.

Fixed this by:

1. Setting the node opacities of outgoing nodes back
to their original values after the outgoing transitions
are complete.

2. Eliminate the initialOpacities map, as this could also
cause memory leaks, in favor of maps that are copied
into the onFinish callbacks. We can do this because
#1 means we no longer have to preserve opacities
across transitions (e.g. from an outgoing to an incoming)