VIRO-852: scene navigation Android crash
We were using a raw pointer for VROSceneController
and wrapping this in a shared pointer when passing it
off to the renderer. So when the renderer stopped using
that scene controller, the shared_ptr would disappear
and the underlying scene controller would be deleted,
leading to crashes when we pushed then popped a
scene.

(Note there's still another, non-native, crash occurring with
Android scene controllers)