VIRO-873: spatial sound bugs
Spatial sounds were not working correctly
with looping and pausing.

1. In order to resume a paused GVR sound,
you have to invoke Resume (invoking Play
 again has no effect)

2. When a GVR sound finishes, it becomes
invalid, and needs to be reloaded and Play
must be invoked again.

3. We need to track whether the sound is
paused in order to know what to do when the
loop is flipped (if the sound is paused, changing
loop to true should not start the sound; otherwise,
it should)