VIRO-877: ensure all sound properties passed to player
Some sound properties were getting lost, as if the
underlying AVAudioPlayer was constructed after
setters were invoked on the VROAudioPlayeriOS,
those settings would not be transferred over to the
underlying player.

Also ensure that the player starts up again if loop
is flipped to true when the player is not paused.