VIRO-777: hierarchical rendering order conflicts
Hierarchical rendering subtrees in the graph were
sorting incorrectly with elements that were not
hierarchical. In the bug given, this caused text to
be rendered before a large flex-box subgraph.

To fix this, we ensure that we only take hierarchy
depth into our rendering sort order computation for
elements that are *within the same hierarchy*. And
then we only compare depths within that hierarchy,
as opposed to overall scene graph depth.

This change also introduces some useful logging for
debugging render order issues.