VIRO-722, 698, 810: reticle bugs
Fixes issues with reticle popping and investigates
background distance for the reticle.

1. We now only shift the reticle's depth when the
user is looking at a foreground object. If the user
looks *off* this foreground object onto the background,
we maintain the reticle's depth. This reduces
popping and works well since the background has a
wide range of acceptable reticle depths, and the eye
tends to track at the last foreground depth anyway.

2. Modified the initial reticle background depth from
5 to 8, after introducing a test loop that shifts the
reticle depths. This is the depth we found that feels
best.