VIRO-891: set head pose for spatial sound
Set the head pose in gvr_audio so that the camera's
position is taken into account when computing the
effects of spatial sound.

1. Head pose is set in the VRODriver's new onFrame()
method, which enables the driver to manage per-frame
events for components it manages.

2. Consolidated gvr_audio so that it's managed by the
VRODriver. Previously it was managed partly by the
VROSceneRendererGVR and the driver; it's cleaner to have
this object managed in only place. Furthermore, by
pulling it into the driver, it now works with OVR.

3. Integrate OVR with gvr_audio.