VIRO-895: remote sound load crash
If a remote sound cannot be loaded, this
leads to a crash. This is related to VIRO-902,
but to mitigate the crash, the IO failures from
downloading sounds are now routed to the
dataError() method in the VROSoundDelegate
implementation (instead of crashing).

With VIRO-902 we will further bubble these
errors up into the JS layer.

This change also takes a step toward VIRO-902
by introducing a success flag into the
downloader.