VIRO-892: Android MediaPlayer error states
The Android MediaPlayer is based on a very
brittle and over engineered state machine. This
commit makes it workable by wrapping Google's
state machine with a more resilient one.

Previously, for example, when pause() was invoked
on a MediaPlayer that was not yet started, the
MediaPlayer would enter an error state and be
unrecoverable. This caused VIRO-892 and probably
other bugs. Now, pause() on an unstarted video
just emits a harmless warning.