VIRO-930, VIRO-685: leak and videos not working
Fixes bugs with videos intermittently crashing or not
working in the theater scene, and with videos not
cleaning up when moving scenes.

1. The intermittent failure of videos was because of a
race condition where we would send some commands to
the video player before it was done initializing on the
rendering thread. Worst case this would cause a crash
as we'd access an invalid PersistentRef for the video
texture. Fixed this by streamlining the construction of
video textures and eliminating the chance of a race
condition.

2. Fixed videos leaking by correcting the memory
cycle referenced in VIRO-685.