VIRO-930: migrate to ExoPlayer
Android's MediaPlayer has been somewhat abandoned
by Google, and wouldn't work with videos in our
theater scene. Migrated to Google's open source
ExoPlayer, which is far more powerful and performant.

The aforementioned videos now work, and lag times are
significantly reduced as the loading is asynchronous.

Note that to get sounds to work, the the delegates must
now be set *before* setup() is invoked on sounds. This
required some small changes in the sound code.