VIRO-806: White bar Appears Under App
- The white bar was a result of the activity not resizing the layout views
correctly as a result of hiding the toolbar and entering full screen. Thus,
fix is to refresh the layout by calling requestLayout() and invalidate()
to allow re-measurement and proper placement of the view to occur.

- Fixed issue where state of the toolbar is not restored when we return from
a Viro experience: the renderer hides the toolbar in fullscreen, and it
remains hidden. The tool bar is restored to its previous configuration.

- Also restored orientation to it's previous configuration when leaving
Viro experience back into react (For example, the React application would
be stuck in landscape after existing VR, even though it may have started
in portrait mode).