VIRO-943: nodes in same hierarchy should share camera distance
When sorting, nodes in the same hierarchy must share the
same camera distance. Otherwise, the sort becomes unstable
and we can end up with bugs like this one, where the "price"
label was drawn first even though it's part of a hierarchy and
should have been drawn after its parent.