VIRO-811: Ovr Crashes on Reload
- React did not call the activity's onDestroy lifecycle stage as apart
of its reloading process. As the renderer's lifecycle has been strongly
tied into the activity's one, it wasn't cleaned up or destroyed properly
on reload. As such when React recreats it again upon reload, the ovr api
detects the incorrect cleanup and crashes. Thus, tied the renderer's lifecycle
into the surface's lifecycle, where if removed and destroyed, also destroys
and cleansup the activity properly on a reload.

- Also fixed toolbar-fullscreen layout issue with ovr.