VIRO-964,VIRO-978: Pause, Play, Loop and Mute Video Bugs
- The video player was calling seekToDefaultPosition whenever play()
is called, causing the video to restart unncessarily. Thus, moved the
setting of seekToDefaultPosition; we only restart the video as needed,
that is, after preparing a video file to be played, or toggling loop
after a video has finished playing.

- Fixed an issue with the Mute property, where if volume was set after
setting mute, it would overrite the mute property incorrectly. As such,
add mute as a separate property in video player to preserve mute states.