VIRO-922 [memory]: pass weak_ptrs to renderer dispatch
The scene tree was not being destroyed even when
the renderer was destroyed because the ViroGVRLayout
was holding onto Runnables to be run in the rendering
thread. These runnables in turn held onto shared_ptrs
to nodes and other components of the scene graph,
causing a leak.

Fixed by using weak_ptrs for all crossing over the thread
boundary via VRODispatchRendererAsync.