VIRO-1014: Ovr Video doesn't Pause when activity is paused.
- The issue was even though we asynchrnously paused our videos during
an Activity's on Pause, the renderer stopped processing all async callbacks
before the onPause and thus did not process the video's onpause request. The
solution is to always process the last java frame before pausing, such that
any tasks left in queue will be flushed and executed. This will thereby
effectively pause the video + audio.