VIRO-992: GVR crash on init
GVR was crashing with an internal stack trace
we couldn’t read, about 20% of the time when doing
a reload. This fix appears to eliminate the problem, by
fixing some thread violations: onPause() and onResume()
needs to occur on the rendering thread.

Since I don’t understand the mechanics of the original
error, can’t be entirely sure whether this fixes the
problem or just reduces its frequency.

Verified that the build no longer crashes on reload,
in my testing, and also verified that the shared_ptr
with dispatch is not keeping the scene renderer from
deallocating.