VIRO-1011, 1012: lighting UBOs corrupted
Fixed two issues with lighting UBOs:

1. They were using an incrementing binding point.
This was a misunderstanding, binding points do not
need to increment, they just need to be unique
per shader. Since we only use one binding point per
shader, it can be fixed at zero. Previously we were
incrementing the binding point until it got to an
unsupported value. This resulted in undefined
behavior.

2. Lighting UBOs were leaking. We kept them as
strong refs in a map, and never removed them when
they were no longer used. Instead, we now hold
strong refs to them in the materials that use them,
and a weak reference in the driver map, so they
are properly disposed.