VIRO-1036: lighting UBO binding point issues
A recent changed switched to using a static binding
point, because we were running out of binding
points if we changed lights often. However, this
doesn’t work either as it turns out binding points
have to be unique per each UBO, as they represent
where a UBO is bound in the EGL context, similar
to texture units.

With this change, instead of using an incrementing
binding point or a static binding point, we use
a managed binding point, utilizing the OpenGL
driver to generate them. When a UBO is destroyed,
it returns its binding point index to the driver for
re-use by another UBO.

This fixes the named bug, where spotlight
angles weren’t working. It turns out they were
not working because the spotlight itself was
being rendered as an omni light (which was part
of a different UBO), because they shared the
same binding point.

To clarify such errors in the future, the lighting
test has also been updated (in the bridge).