VIRO-1017: GL objects deleted in wrong EGL context
Animated objects were randomly disappearing after
reloads. This was because those objects’ vertex arrays
(GL objects) backing them were being deleted. This
would happen because of the following sequence of
events:

1. System allocates vertex array X in EGL Context 1
2. Object is animated
3. Reload occurs, EGL Context 1 is destroyed, and
    EGL Context 2 is created
4. The old object is not yet deleted because its animation,
    though not visible, is still executing
5. A new object is created using the same vertex array
    ID, X, as the still-not-deleted object in the previous
    context.
6. The animation finishes, and the system deletes
    vertex array X. However, this deletes vertex array
    X not in the old context (where it originated) but in
    the *new* context, in which it represents an entirely
    different object.

To protect against this, we must ensure that we only
delete GL objects that are part of the *active* EGL context.
Our VRODriver objects are a good proxy for this, they hold
all context-specific scope already, and are destroyed when
the context is destroyed.

To fix this bug then, we ensure that every object that owns
a GL object maintains a weak reference to its context’s
driver. When it comes time to delete the object, we check if
the driver is still alive. If it is, it’s safe to delete. Otherwise we
do not delete (and we don’t need to, because when a context
is destroyed its underlying graphics objects are always destroyed
along with it).

This *vast majority* of this change was, unfortunately, ensuring
that VRODriver uses shared_ptr everywhere. We were somewhat
haphazardly sometimes using it as a shared_ptr, sometimes as a
ref, and sometimes as a raw pointer. Now we use it as a shared_ptr
everywhere except the process-intensive core rendering cycle,
where shared_ptr copies are too expensive.

The real focus of this change is small, in the following files where we
perform the check of the driver before deleting:

1. VROGeometrySubstrateOpenGL
2. VROTextureSubstrateOpenGL
3. VROLightingUBO

The rest is plumbing.