VIRO-1059: memory corruption loading invalid materials
The random ‘malloc’ errors in the heart scene were
happening when we reloaded materials, invalidating
the prior ones, and then asynchronously finished
loading the Heart OBJ and tried to set those invalid
materials onto its geometry.

Fixed by making the setMaterials function return false
if any materials have been destroyed. The bridge can
then respond as it sees fit (in this case it ignores the
failure, as the failure is indicative of a stale OBJ
being loaded).