VIRO-1051: render order issues interleaving hierarchies
The way the render order comparison was written, with an
if statement, made it possible for objects within different
hierarchies to be compared to one another, resulting in an
invalid final rendering order because not every object
*within* one hierarchy would be compared to one another.

For example, if we had hierarchy 0 with objects A and B and
hierarchy 1 with objects C and D, the sort function might
compare A and C and B and C, discover that A < C and
C < B, and deduce therefore that A < B. It would have made
this comparison without actually comparing A
and B.

This would be wrong because our if statement made our
ordering inconsistent. Because of these problems we’ve
eliminated the if statement here and use one completely
consistent sort ordering.