VIRO-13: display FPS with rapidly changing text
Commit on the renderer side to display FPS.
The following changes were necessary:

1. Add a DebugHUD that can be used to display
debug information, and integrate with render
cycle.

2. Allow typefaces to preload and cache glyphs,
so that we can change the VROText displaying
the FPS faster. This is also a requirement because
we cannot load new glyphs in the midst of the
render cycle, since loading glyphs takes over the
EGL context.

3. Expose functionality to enable and disable the
Debug HUD.