VIRO-1117 [surface modifiers]: modify iOS shaders
Modify iOS shaders to support surface modifiers
in the fragment shader. We essentially pipe
all the modifiable uniforms into a _surface struct,
which modifiers will be able to modify.

This also simplifies the shader code in general,
in that all shaders now share the same surface
struct instead of having separate structs for each
lighting model.