VIRO-1090: thread restriction framework
Created  new class, VROThreadRestricted, that allows
us to restrict a class to a specified thread (either the
thread of the constructor or a manually specified thread).
Subclasses can then easily assert they’re on the correct
thread.

Apply this to both VRONode and VROScene, which are
primary touchpoints for rendering object access. We don’t
want to overuse this and apply it to every rendering object
subclass, because it does have a memory/perf cost, and
we can capture the vast majority of violations by checking
these top-level rendering objects.