VIRO-1090: thread restriction fixes
Made a slew of thread restriction fixes for Android,
ensuring that nodes (primarily) are only accessed
via the rendering thread.

1. Made VROInputPresenter lazily loaded on the
rendering thread, so that its rootNode is only constructed
on the rendering thread. To do this, had to remove its
reliance on the VRORenderContext, which shouldn’t
usually be long-held anyway. Made this change by
adding the VROCamera to the hit test results, eliminating
the need for a context in the presenters.

2. Cleaned up hit test related functions while doing this,
ensuring the parameter ordering is consistent.

3. Removed shared_ptr from input controller base as well.

4. Fixed debug HUD so its node is constructed in its
rendering init function.

5. Added more debug information to thread restricted.

6. Allow the thread ID to be set after initialization; there
was no other way to grab the rendering thread ID when
creating nodes, so we have to set the ID immediately
after construction.