VIRO-1090, 924: threading with OBJ loading
Asynchronously copy resources for OBJ and
ensure there are no thread violations. The key
change here is the VRONode for the object
is no longer returned when the OBJ loading
is initialized; instead it’s bubbled up through
the delegates.

Also removed from booleans to reduce the
amount of state in ObjectJni; it’s simpler now.