VIRO-918 [renderer]: preserve rendering tree when animate out
There were two issues that were causing the rendering
tree to be deconstructed on Android before a scene
could transition out. The first issue is fixed here: we were
removing children from their parent upon destruction of the
PersistentRef. There’s no reason to do this, and doing so
was destroying the rendering tree the minute the bridge
tree was torn down.

Fixed and checked to ensure we’re not leaking memory on
scene transitions and reloads.