[shaders]: unify constant and textured shaders
Use shader modifiers to unify the constant and
textured shader for each lighting model. This
substantially reduces code bloat and the number
of shaders we have to manager.

To make this work we also had to change shader
modifiers so that *multiple* modifiers can be
applied to each entry point and section. Achieved
this by appending each modifier with the pragma
directive that defines where the next modifier
can slip in.