VIRO-40 [renderer]: normal maps
Add support for normal mapping in the renderer. This
required the following changes, in this commit:

1. Simplify geometry source creation by delegating it
to VROShapeUtils. Previously we had a lot of repeated
code creating the same geometry source for the same
vertex formats, over and over. Now it’s all consolidated
in VROShapeUtils. This enables us to modify the vertex
layout without having to change every shape.

2. Make OBJ files use VROShapeVertexLayout, instead of
their own buffer. This pulls them into line with every other
object we render.

3. Add a vec4 tangent attribute to our vertex layout. The
tangent is necessary to render normap maps.

4. Add function to compute the tangent for each vertex
in a shape. This is in VROShapeUtils. It computes both
the tangent and the handedness, so that we can compute
the bitangent in the shader.

5. Compute the TBN matrix in the vertex shader. The
TBN matrix converts from tangent space (the space
of the normal map) into world space. We now compute
this matrix instead of simply just a normal vector in the
vertex shader.

6. Create a shader modifier that samples from the
normal map and multiplies by the TBN matrix. With the
new simplified shader system, we can perform the
fragment shader normal mapping with just a simple shader
modifier.