VIRO-441: use shared_ptr for driver in render-cycle
We were using shared_ptr everywhere else
for the driver; this finalizes it and uses it in the
render-cycle too. To avoid expensive shared_ptr
copies (which are expensive because they’re
thread-safe) we instead pass references to the
shared_ptr. These are safe because the calling
function owns a shared_ptr that stays alive.

We need this change so that we can properly
do VIRO-441, asynchronous texture loading. The
Texture class will be need to invoke a lambda,
and to do that safely we need a weak_ptr to
the driver (hence we need a shared_ptr to the
driver).