VIRO-441: async texture loading
Introduces the VROFrameScheduler, which can
be used to schedule code that must be run on the
rendering thread (with a valid EGL context), in a
time-limited queue. This queue ensures that the
code is only run if there are spare cycles during
a frame, to limit stutters in the frame-rate.

The first use of the scheduler, here, is for loading
textures asynchronously.