VIRO-441: preload image handling on Android
Android did not have a ‘blank’ preload image, this
was a stray TODO that was causing a crash.

Fix this for Android, and change the blank image from
white to transparent, so that it matches the default
placeholder image used on the bridge. This should also
fix any bugs related to white “flashes” on iOS before
videos load.