VIRO-1155: empty text was crashing
When in justify mode, we’d crash if there was
no space in the text. When justifying we’d try
to get the length of the space character, using
the glyph map from the text: if the text had no
space, we would crash. Fixed so that space
is always included in the glyph map.