VIRO-1153: move from modifier hash to unique key
The shader modifier hash being used would fail
after a few hundred shader modifiers were created
(which is what happens with onFuse). This would
cause a crash as the system would think one shader
modifier batch was another, and uniforms would not
be found.

Added protection against this through assertions
to detect early, and replaced the shader modifier
hash with a unique modifier key.

This does not fix the shader and shader modifier
memory leak.