VIRO-1153: memory leaks from onFuse and driver
onFuse exposed memory leaks because it creates
hundreds of new shaders, and we don’t have any
system for purging shaders. Introduced a new time-based
shader purging process in VRODriver. This can be
extended later to more resources, if required.

Also simplified the uniform loading process in
VROMaterialSubstrateOpenGL, as there was quite a bit
of code duplication.