Support named thread restrictions
When changing orientations we were running into
crashes if the rendering thread changed when the
surface was replaced. Fix this by allowing “named”
threads in the ThreadRestriction class. Objects can
restrict themselves to the “Renderer” thread for
example, and on startup the system has to manually
associate a given thread with the “Renderer” thread.