VIRO-1108: Daydream reticle stablization issues
This cr fixes two main issues with the daydream reticle. Firstly, it addresses
the  was snapping / jumping that occurs as the user points onto objects.
Secondly, it also fixes the "miss-clicked" issue where the reticle
is actually behind the view, instead of on it, resulting in a mis-click.

The fix is to perform hit tests shotting  a vector from the user's observation
point (the camera) instead of the controller's observation point. This is
done by projecting the controller's forward vector onto a point in the background,
then getting the user's forward vector to that point, and then performing the
hit test with that new forward vector.

In addition, this cr also disables fuse by default (it wasn't before), and
re-adjusts the size of the reticle to better reflect the one on the daydream
homepage.