VIRO-1157: dynamic far clipping plane
Compute the FCP each frame as the distance of
the furthest object from the camera. More
specifically, we use the farthest point on each object’s
bounding box from the camera.

This fixes the theater clipping issue, where the edges
of the video screen were clipped because we were
using a fixed FCP that wasn’t distance enough.