VIRO-858: mutable geometry + 6-material boxes
This commit enables 6-material boxes (one material
per surface) but in order to do this introduces two other
features:

1. Mutable geometries. There is now an easy pathway
to make a geometry mutable, by changing its vertex
layout and updating its substrate. This was necessary
because we have to mutate the box to change its
geometry elements if the box switches from 1 material
to 6. This commit also demonstrates this feature by
adding animatable setWidth/setHeight/setLength
methods to VROBox.

2. Mutate geometries in response to materials.
Previously, materials were mutated by grabbing the
geometry array and directly manipulating it. That
pathway has been closed off; now the only way to
change materials is to invoke setMaterials() in
VROGeometry. This enables VROBox to override
setMaterials to mutate the geometry in response to
 the number of materials received.